move BaseError, subclass it with InternalError and UserError
