+++ data-processed_id2iri_mapping_2023-03-10_142838.json
@@ -1,0 +1,13 @@
+{
+    "Anubis.jpg_daa1164a-5cdb-4010-ab90-81f8e36168cb": "http://rdfh.ch/0123/jTcxLg0WRTq6IBiHxSvOAA",
+    "Basho_Horohoroto.jpg_dce0b6ec-c2ad-4911-9460-7d5f7c769ada": "http://rdfh.ch/0123/Ahf-Y57gT3eKf7vy_Ct62Q",
+    "BM1888-0601-716.png_036b5fcd-ea4d-4e8a-ba66-5e5c215e701a": "http://rdfh.ch/0123/G5gIFXiYT9OUpJl-e6V9eQ",
+    "GibeonMeteorite.jpg_bcabb44a-0bfd-4a53-816e-01c9fc59dd11": "http://rdfh.ch/0123/0SvGMRz-TI--dDZDibXwIg",
+    "Anubis_408c045e-d9c0-47fb-86b3-40d43b987f56": "http://rdfh.ch/0123/oJxNohLAQSenLV-i-zkG4g",
+    "Meteorite_0527463f-a9df-4760-bbfa-3ac651281b38": "http://rdfh.ch/0123/_5tMNsfnRgGFCsM89cC8sg",
+    "BM1888-0601-716_6c681156-9f7d-47c7-9a23-0ae865526200": "http://rdfh.ch/0123/NE7RIzIkRo-jpEeslzUmBA",
+    "Horohoroto_a3e77c8c-019d-4b91-8ff8-52a01c405d98": "http://rdfh.ch/0123/K-SN7xqcRO-rOKw9XeM-9g",
+    "annotation_to_anubis": "http://rdfh.ch/0123/17pXYvVRQUCUfz35hKEFRQ",
+    "region_of_meteorite": "http://rdfh.ch/0123/4Mi7P8RLTcyCvIEgzyCHOQ",
+    "link_BM1888-0601-716_horohoroto": "http://rdfh.ch/0123/2LyBGvjyT3G1VmDa7Yb7lA"
+}
