+++ src/dsp_tools/models/group.py
@@ -8,7 +8,8 @@ from pystrict import strict
 
 from dsp_tools.models.langstring import LangString
 from dsp_tools.models.connection import Connection
-from dsp_tools.models.helpers import Actions, BaseError
+from dsp_tools.models.helpers import Actions
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.models.model import Model
 from dsp_tools.models.project import Project
