+++ src/dsp_tools/models/helpers.py
@@ -6,6 +6,7 @@ from typing import NewType, Optional, Any, Tuple, Union, Pattern
 
 from pystrict import strict
 
+from dsp_tools.models.exceptions import BaseError
 
 #
 # here we do some data typing that should help
@@ -40,33 +41,6 @@ class IriTest:
     def test(cls, val: str) -> bool:
         m = cls.__iri_regexp.match(val)
         return m.span()[1] == len(val) if m else False
-
-
-@strict
-class BaseError(Exception):
-    """
-    A basic error class
-    """
-    _message: str
-
-    def __init__(self, message: str) -> None:
-        """
-        Constructor for error message
-        :param message: error message string
-        """
-        super().__init__()
-        self._message = message
-
-    def __str__(self) -> str:
-        """
-        Convert to string
-        :return: stringyfied error message
-        """
-        return self._message
-
-    @property
-    def message(self) -> str:
-        return self._message
 
 
 @unique
