+++ src/dsp_tools/models/model.py
@@ -1,5 +1,5 @@
 from dsp_tools.models.connection import Connection
-from dsp_tools.models.helpers import BaseError
+from dsp_tools.models.exceptions import BaseError
 
 
 class Model:
