+++ src/dsp_tools/models/project.py
@@ -9,7 +9,8 @@ from pystrict import strict
 
 from dsp_tools.models.set_encoder import SetEncoder
 from dsp_tools.models.connection import Connection
-from dsp_tools.models.helpers import Actions, BaseError
+from dsp_tools.models.helpers import Actions
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.models.langstring import Languages, LangString
 from dsp_tools.models.model import Model
