+++ src/dsp_tools/models/projectContext.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from dsp_tools.models.connection import Connection
 from dsp_tools.models.group import Group
-from dsp_tools.models.helpers import BaseError
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.models.project import Project
