+++ src/dsp_tools/utils/excel_to_json_project.py
@@ -2,7 +2,7 @@ import json
 import os
 import re
 
-from dsp_tools.models.helpers import BaseError
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.utils.excel_to_json_lists import excel2lists
 from dsp_tools.utils.excel_to_json_properties import excel2properties
 from dsp_tools.utils.excel_to_json_resources import excel2resources
