+++ src/dsp_tools/utils/excel_to_json_resources.py
@@ -8,7 +8,7 @@ import jsonpath_ng.ext
 import pandas as pd
 import regex
 
-from dsp_tools.models.helpers import BaseError
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.utils.shared import prepare_dataframe, check_notna
 
 languages = ["en", "de", "fr", "it", "rm"]
