+++ test/unittests/test_excel_to_json_lists.py
@@ -10,7 +10,7 @@ import jsonpath_ng.ext
 import pandas as pd
 import regex
 
-from dsp_tools.models.helpers import BaseError
+from dsp_tools.models.exceptions import BaseError
 from dsp_tools.utils import excel_to_json_lists as e2l
